menyelesaikan halaman daftar buku & menambahkan halaman data petugas

+++ app/Http/Controllers/AdminController.php
@@ -6,9 +6,13 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Models\Buku_model;
 use App\Models\PetugasModel;
+use App\Models\User;
+use App\Models\RoleUser;
 use DateTime;
 use Illuminate\Support\Facades\File;
 use App\Exports\BukuExport;
+use App\Exports\PetugasExport;
+use Illuminate\Container\RewindableGenerator;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
 
@@ -19,9 +23,15 @@ class AdminController extends Controller
         return view('admin.dashboard');
     }
 
-    public function daftarBuku()
+    public function daftarBuku(Request $request)
     {
-        $buku = Buku_model::all();
+        if($request->cari != '') {
+            $cari = $request->cari;
+            $request->session()->put('cari', $request->cari);
+        }else{
+            $cari = $request->session()->get('cari');
+        }
+        $buku = Buku_model::getBuku($cari);
         return view('admin.daftar-buku', compact('buku'));
     }
 
@@ -32,7 +42,7 @@ class AdminController extends Controller
             'pengarang' => 'required',
             'penerbit' => 'required',
             'tahun_terbit' => 'required',
-            'foto' => 'required|image|mimes:jpeg,png,jpg|max:2048',
+            'foto' => 'image|mimes:jpeg,png,jpg|max:2048',
             'bahasa' => 'required',
             'genre' => 'required',
             'jml_halaman' => 'required',
@@ -43,8 +53,12 @@ class AdminController extends Controller
         $buku = DB::table('buku')->max('id_buku');
         $id_buku = $buku + 1;
         $judul = $request->judul;
-        $imageName = str_replace(" ", "_", $judul) . '.' . $request->foto->extension();
-        $request->foto->move(public_path('img/buku'), $imageName);
+        if($request->foto != '') {
+            $imageName = str_replace(" ", "_", $judul) . '.' . $request->foto->extension();
+            $request->foto->move(public_path('img/buku'), $imageName);
+        } else {
+            $imageName = 'default.jpg';
+        }
         Buku_model::create([
             'id_buku'   => $id_buku,
             'judul' => $judul,
@@ -56,7 +70,8 @@ class AdminController extends Controller
             'genre' => $request->genre,
             'jml_halaman'   => $request->jml_halaman,
             'stok'  => $request->stok,
-            'created_at'    => date('Y-m-d h:i:s')
+            'created_at'    => date('Y-m-d h:i:s'),
+            'updated_at'    => null
         ]);
 
         return redirect('daftarBuku')->with('status', 'Berhasil menambahkan buku baru.');
@@ -66,7 +81,9 @@ class AdminController extends Controller
     {
         foreach ($request->id as $id) {
             $image = Buku_model::where('id_buku', $id)->first();
-            File::delete('img/buku/' . $image->foto);
+            if($image->foto != 'default.jpg') {
+                File::delete('img/buku/' . $image->foto);
+            }
 
             Buku_model::where('id_buku', $id)->delete();
         }
@@ -150,21 +167,65 @@ class AdminController extends Controller
         return redirect('detailBuku/' . $request->id_buku)->with('status', 'Data buku berhasil diubah.');
     }
 
-    // public function exportBukuExcel()
-    // {
-    //     return Excel::download(new BukuExport, 'Daftar-Buku.xlsx');
-    // }
-
-    // public function exportBukuPdf()
-    // {
-    //     $buku = Buku_model::all();
-    //     $pdf = PDF::loadview('admin.export-buku', compact('buku'));
-    //     return $pdf->download('daftar-buku-pdf');
-    // }
-
-    public function dataPetugas()
+    public function exportBukuExcel()
     {
-        $petugas = PetugasModel::all();
+        return Excel::download(new BukuExport, 'Daftar-Buku.xlsx');
+    }   
+
+    public function exportBukuPdf($cari = null)
+    {
+        $buku = Buku_model::getBuku($cari);
+        $pdf = PDF::loadview('admin.export-buku', compact('buku'))->setPaper('a4', 'landscape');
+        return $pdf->download('daftar-buku.pdf');
+    }
+
+    public function refreshBuku()
+    {
+        session()->forget('cari');
+        return redirect('daftarBuku');
+    }
+
+    public function dataPetugas(Request $request)
+    {
+        if($request->cari != '') {
+            $cari = $request->cari;
+            $request->session()->put('cari', $cari);
+        }else{
+            $cari = $request->session()->get('cari');
+        }
+
+        $petugas = PetugasModel::getPetugas($cari);
         return view('admin.data-petugas', compact('petugas'));
+    }
+
+    public function hapusPetugas(Request $request)
+    {
+        foreach ($request->id as $id) {
+            $image = User::where('id', $id)->first();
+            if ($image->image != 'default.png') {
+                File::delete('img/user_img/' . $image->image);
+            }
+
+            PetugasModel::where('id', $id)->delete();
+        }
+        return redirect('dataPetugas')->with('status', 'Data petugas berhasil dihapus.');
+    }
+
+    public function refreshPetugas()
+    {
+        session()->forget('cari');
+        return redirect('/dataPetugas');
+    }
+
+    public function exportPetugasExcel()
+    {
+        return Excel::download(new PetugasExport, 'Data Petugas.xlsx');
+    }
+
+    public function exportPetugasPdf($cari = '')
+    {
+        $petugas = PetugasModel::getPetugas($cari);
+        $pdf = PDF::loadview('admin.export-petugas', compact('petugas'))->setPaper('a4', 'landscape');
+        return $pdf->download('Data Petugas.pdf');
     }
 }
